Updated schema to replace double type with decimal
one more fix

+++ dataflow-website/batch-developer-guides/batch/batchsamples/billrun/src/main/resources/schema.sql
@@ -5,5 +5,5 @@ CREATE TABLE IF NOT EXISTS BILL_STATEMENTS
    last_name varchar(50),
    minutes int,
    data_usage int,
-   bill_amount double
+   bill_amount decimal(10,2)
 );
